Fix some relative imports of ButlerURI now that it moved
